fix: extract shared SummaryStatusIcon, fix approval workflow triggers
- Extract duplicate OverallStatusIcon/GroupStatusIcon into shared
  SummaryStatusIcon component (addresses review comment)
- Remove opened/synchronize/reopened from approval workflow triggers
  so it only runs on review or label events (fixes false failures)
